+++ cardinal-components-item/src/main/java/dev/onyxstudios/cca/internal/item/StaticItemComponentPlugin.java
@@ -26,10 +26,8 @@ import com.google.common.collect.Iterables;
 import dev.onyxstudios.cca.api.v3.component.ComponentContainer;
 import dev.onyxstudios.cca.api.v3.component.ComponentKey;
 import dev.onyxstudios.cca.api.v3.component.ComponentV3;
-import dev.onyxstudios.cca.api.v3.item.ItemComponentFactory;
-import dev.onyxstudios.cca.api.v3.item.ItemComponentFactoryRegistry;
-import dev.onyxstudios.cca.api.v3.item.ItemComponentFactoryV2;
-import dev.onyxstudios.cca.api.v3.item.ItemComponentInitializer;
+import dev.onyxstudios.cca.api.v3.component.TransientComponent;
+import dev.onyxstudios.cca.api.v3.item.*;
 import dev.onyxstudios.cca.internal.base.LazyDispatcher;
 import dev.onyxstudios.cca.internal.base.asm.CcaAsmHelper;
 import dev.onyxstudios.cca.internal.base.asm.StaticComponentLoadingException;
@@ -131,6 +129,26 @@ public final class StaticItemComponentPlugin extends LazyDispatcher implements I
     public <C extends Component> void registerFor(Identifier itemId, ComponentKey<C> type, ItemComponentFactory<? extends C> factory) {
         Objects.requireNonNull(itemId);
         this.register(itemId, type, factory);
+    }
+
+    @Override
+    public <C extends ItemComponent> void register(Predicate<Item> test, ComponentKey<? super C> type, ItemComponentFactory<C> factory) {
+        this.registerFor(test, type, ItemComponent.wrapFactory(type, factory));
+    }
+
+    @Override
+    public <C extends ItemComponent> void register(Item item, ComponentKey<? super C> type, ItemComponentFactory<C> factory) {
+        this.registerFor(item, type, ItemComponent.wrapFactory(type, factory));
+    }
+
+    @Override
+    public <C extends TransientComponent> void registerTransient(Predicate<Item> test, ComponentKey<? super C> type, ItemComponentFactory<C> factory) {
+        this.registerFor(test, type, factory);
+    }
+
+    @Override
+    public <C extends TransientComponent> void registerTransient(Item item, ComponentKey<? super C> type, ItemComponentFactory<C> factory) {
+        this.registerFor(item, type, factory);
     }
 
     @Override