+++ changelog.md
@@ -1,4 +1,46 @@
 ------------------------------------------------------
+Version 2.7.10
+------------------------------------------------------
+Additions
+- Added `ItemComponent`, an experimental component implementation that stores all its data in the stack's tag
+- Added `ScoreboardComponentFactoryV2`, giving access to the server holding the scoreboard (if any)
+- Added `TeamComponentFactoryV2`, giving access to the scoreboard holding the team and its server (if any)
+
+Changes
+- Promoted `ComponentPacketWriter` and `PlayerSyncPredicate` to stable API
+- Further optimized ItemStack components
+- A copy of the MIT license is now bundled with every CCA jar
+
+------------------------------------------------------
+Version 2.7.9
+------------------------------------------------------
+- Mcdev annotations are no longer a transitive dependency
+
+------------------------------------------------------
+Version 2.7.8
+------------------------------------------------------
+Changes
+- Promoted `ComponentKey#sync(provider, writer)`, `ComponentKey#sync(provider, writer, predicate)`,
+ `ComponentKey#isProvidedBy`, and `ComponentKey#getNullable` to stable API
+
+Fixes
+- Fixed backward compatibility with SyncedComponent
+
+------------------------------------------------------
+Version 2.7.7
+------------------------------------------------------
+Fixes
+- Fixed NBT deserialization not respecting component registration order
+
+------------------------------------------------------
+Version 2.7.6
+------------------------------------------------------
+Changes
+- Optimized ItemStack component initialization
+- Promoted `ComponentKey#syncWith` to (experimental) public API
+- Promoted `ClientTickingComponent`, `ServerTickingComponent` and `CommonTickingComponent` to stable API
+
+------------------------------------------------------
 Version 2.7.5
 ------------------------------------------------------
 Additions